Crate icu_datagen

icu_datagen contains command-line tools to generate and process ICU4X data.

The tools include:

icu4x-datagen: Read source data (CLDR JSON) and dump ICU4X-format data.

icu4x-testdata-download: Download fresh CLDR JSON for testdata.

icu4x-key-extract: Extract ResourceKey objects present in a compiled executable.

More details on each tool can be found by running --help.

Examples

Generate ICU4X Postcard blob (single file) for all keys and all locales:

$ cargo run --bin icu4x-datagen -- \
    --cldr-tag 39.0.0 \
    --all-keys \
    --all-locales \
    --format blob \
    --out /tmp/icu4x_data/icu4x_data.postcard

Extract the keys used by an executable into a key file:

$ cargo build --example work_log --release --features serde
$ cargo make icu4x-key-extract \
    target/release/examples/work_log \
    /tmp/icu4x_data/work_log+keys.txt
$ cat /tmp/icu4x_data/work_log+keys.txt

Generate ICU4X JSON file tree from the key file for Spanish and German:

$ cargo run --bin icu4x-datagen -- \
    --cldr-tag 39.0.0 \
    --key-file /tmp/icu4x_data/work_log+keys.txt \
    --locales es \
    --locales de \
    --out /tmp/icu4x_data/work_log_json

Modules

cldr

icu_datagen::cldr contains implementations of the ICU4X data provider interface based on the JSON files shipped by CLDR. Create a CldrPaths and then pass it into create_exportable_provider.

segmenter

icu_datagen::segmenter contains implementations of the ICU4X data provider interface based on Unicode properties and TOML files implementing Unicode Standard Annex #14 and Unicode Standard Annex #29 breaking rules.

uprops

icu_datagen::uprops contains implementations of the ICU4X data provider interface backed by TOML files exported by the ICU4C icuwriteuprops tool. Create a directory containing TOML files for the necessary Unicode properties and then pass the path to the desired provider.

Functions

get_all_keys

List of all supported keys

get_registry

Get a registry that has the appropriate ConvertData and IterableDynProvider implementations