Function icu_datagen::datagen

pub fn datagen(
    locales: Option<&[LanguageIdentifier]>, 
    keys: &[ResourceKey], 
    sources: &SourceData, 
    outs: Vec<Out>
) -> Result<(), DataError>

Runs ICU4X datagen.

The argument are used as follows:

locales: If this is present, only locales that are either und or contained (strictly, i.e. en != en-US) in the slice will be generated. Otherwise, all locales supported by the source data will be generated.

keys: The keys for which to generate data. See all_keys, keys, keys_from_file, keys_from_bin.

sources: The underlying source data. CLDR and/or ICU data can be missing if no requested key requires them, otherwise MISSING_CLDR_ERROR or MISSING_ICUEXPORT_ERROR will be returned.

out: The output format and location. See the documentation on Out