Macro icu_datagen::create_datagen_provider

macro_rules! create_datagen_provider {
    ($source_data:expr) => { ... };
    ($source_data:expr, [ $($constructor:path),+, ]) => { ... };
}

Create a data provider reading from source files that generates data for all, or a subset, of ICU4X.

The macro behaves like a function that takes the following arguments:

An instance of SourceData (required)

A list of providers to instantiate (optional)

The return value is a complex type that implements all of the key data provider traits.

To create a data provider for all of ICU4X:

use icu_datagen::SourceData;

// This data provider supports all keys required by ICU4X.
let provider = icu_datagen::create_datagen_provider!(SourceData::default());

To create a data provider for a subset:

use icu_datagen::SourceData;

// This data provider supports the keys for LocaleCanonicalizer.
let provider = icu_datagen::create_datagen_provider!(
    SourceData::default(),
    [
        icu_datagen::transform::cldr::AliasesProvider,
        icu_datagen::transform::cldr::LikelySubtagsProvider,
    ]
);