Module icu_datagen::segmenter

icu_datagen::segmenter contains implementations of the ICU4X data provider interface based on Unicode properties and TOML files implementing Unicode Standard Annex #14 and Unicode Standard Annex #29 breaking rules.

Important: This data provider implementation is not optimized for production use. Read more in the data provider docs.

Examples

use icu_datagen::segmenter::SegmenterRuleProvider;
let provider = SegmenterRuleProvider::try_new(
    "/path/to/segmenter/data/directory",
    "/path/to/uprops/data/directory",
)
.expect_err("Specify a real directory in the line above");

Exporting data

To generate the data required by the segmenters, run cargo make testdata from the top level.

Structs

SegmenterRuleProvider

A data provider reading from segmenter rule files.

Functions

create_exportable_provider

segmenter_data_root

Returns the absolute path to the directory containing the segmenter raw data.