Struct icu_datagen::uprops::EnumeratedPropertyCodePointTrieProvider

pub struct EnumeratedPropertyCodePointTrieProvider { /* private fields */ }

This data provider returns CodePointTrie data inside a UnicodePropertyMap data struct. The source data is the same as that of the other properties providers, which is a TOML file of data for the property(-ies) desired, as given by the ICU4C property data exporter tool.