Crate icu_datagen

icu_datagen is a library to generate data files that can be used in ICU4X data providers.

Data files can be generated either programmatically (i.e. in build.rs), or through a command-line utility.

Examples

build.rs

use icu_datagen::*;
use icu_locid::langid;
use std::fs::File;
use std::path::PathBuf;

fn main() {
    icu_datagen::datagen(
        Some(&[langid!("de"), langid!("en-AU")]),
        &icu_datagen::keys(&["list/and@1"]),
        &SourceData::default().with_uprops(PathBuf::from("/path/to/uprops/root")),
        Out::Blob(Box::new(File::create("data.postcard").unwrap())),
        false,
    ).unwrap();
}

Command line

The command line interface is available with the bin feature.

cargo run --features bin -- \
    --uprops-root /path/to/uprops/root \
    --all-keys \
    --locales de,en-AU \
    --format blob \
    --out data.postcard

More details can be found by running --help.

Modules

transform

Exposes all available data transfomers

Macros

create_datagen_provider

Create a data provider reading from source files that generates data for all, or a subset, of ICU4X.

Structs

SourceData

Bag of options for datagen source data.

Enums

Out

The output format.

Functions

datagen

Runs ICU4X datagen.

get_all_keys

List of all supported keys

is_missing_cldr_error

Identifies errors that are due to missing CLDR data.

is_missing_uprops_error

Identifies errors that are due to missing Unicode properties data.

keys

Parses a list of human-readable key identifiers and returns a list of ResourceKeys. Invalid key names are ignored.

keys_from_file

Parses a file of human-readable key identifiers and returns a list of ResourceKeys. Invalid key names are ignored.